navbar added, reducer separated, cosmetic changes
