+++ task-01/src/containers/Form.js
@@ -1,21 +1,6 @@
 import React, { useState, useReducer, useEffect } from "react";
-
-function reducer(state, action){
-    switch (action.type) {
-      case "fieldChange": {
-        const newState = {...state, [String(action.payload.name)]: String(action.payload.value)}
-        return newState;
-      }
-
-      case "setData": {
-          const newState = action.payload;
-          return newState;
-      }
-      default: {
-        return {...state};
-      }
-    }
-}
+import { reducer } from "../libs/reducer";
+import './Form.css';
 
 export default function Form(){
   const [userData, dispatch] = useReducer(reducer, {});
@@ -33,7 +18,6 @@ export default function Form(){
       }
     })
     .then(data => {
-      console.log("data: ", data);
       if(data){
         let d = data[0];
         dispatch({type: "setData", payload: d});
@@ -74,29 +58,27 @@ export default function Form(){
   }
   
     return ( isloading ? <div>Loading...</div> : (
-      <div>
-          <form onSubmit={handleSubmit}>
-            <label>
-                First Name:   
-                <input type="text" id="firstName" defaultValue={userData.firstName} onChange={handleChange}/>
-            </label><br />
-            <label>
-                Last Name:   
-                <input type="text" id="lastName" defaultValue={userData.lastName} onChange={handleChange}/>
-            </label><br />
-            <label>
-                Birth Date:    
-                <input type="text" id="birthDate" defaultValue={userData.birthDate} onChange={handleChange}/>
-            </label><br />
-            <label>
-                Eye Color:     
-                <input type="text" id="eyeColor" defaultValue={userData.eyeColor} onChange={handleChange}/>
-            </label><br />
-            <input type="submit" value="submit"/><br />
-            <p>{}</p>
-        </form>
-            <button onClick={() => console.log("userdata from click",userData)}>log userData</button>
-      </div>
+        <div>
+            <form onSubmit={handleSubmit}>
+                <div className="fields">
+                    <div className="labels">
+                        <label>First Name:</label><br />
+                        <label>Last Name:</label><br />
+                        <label>Birth Date:</label><br />
+                        <label>Eye Color:</label><br />
+                    </div>
+                    <div className="inputs">
+                        <input type="text" id="firstName" defaultValue={userData.firstName} onChange={handleChange}/><br />
+                        <input type="text" id="lastName" defaultValue={userData.lastName} onChange={handleChange}/><br />
+                        <input type="text" id="birthDate" defaultValue={userData.birthDate} onChange={handleChange}/><br />
+                        <input type="text" id="eyeColor" defaultValue={userData.eyeColor} onChange={handleChange}/><br />
+                    </div>
+                </div>
+                <input type="submit" value="submit"/><br />
+                <p>{}</p>
+            </form>
+                <button onClick={() => console.log("userdata from click",userData)}>log userData</button>
+        </div>
     )
     )
