+++ task-01/src/containers/NavBar.css
@@ -1,0 +1,12 @@
+.navBar {
+    width: 100%;
+    padding: 0;
+    height: 50px;
+    position: relative;
+}
+
+.nav {
+    float: right;
+    padding: 5px;
+    margin: 10px;
+}
